initial refactor complete
