+++ mypy.ini
@@ -138,12 +138,6 @@ disallow_untyped_calls = False
 [mypy-wandb.sdk.wandb_settings]
 disallow_untyped_calls = False
 
-[mypy-wandb.sdk.data_types]
-disallow_untyped_calls = False
-
-[mypy-wandb.sdk.data_types._dtypes]
-ignore_errors = True
-
 [mypy-wandb.sdk.wandb_artifacts]
 disallow_untyped_calls = False
 
@@ -157,6 +151,16 @@ ignore_errors = True
 ignore_errors = True
 
 [mypy-wandb.sdk.wandb_save]
+ignore_errors = True
+
+# -----------------------------
+# wandb/sdk/data_types/
+# -----------------------------
+
+[mypy-wandb.sdk.data_types.*]
+disallow_untyped_calls = False
+
+[mypy-wandb.sdk.data_types._dtypes]
 ignore_errors = True
 
 # -----------------------------
