+++ wandb/sdk/data_types/__init__.py
@@ -17,8 +17,6 @@ from .helper_types.Classes import Classes
 
 # TODO: Remove this once /data_types are all moved in this dir.
 from .base_types.Media import _numpy_arrays_to_lists
-# TODO: remove once refactored
-from .private._common import MEDIA_TMP, get_max_cli_version , is_numpy_array
 
 import logging
 import re
