+++ wandb/sdk/data_types/_private.py
@@ -1,0 +1,3 @@
+import tempfile
+
+MEDIA_TMP = tempfile.TemporaryDirectory("wandb-media")
