+++ wandb/sdk/data_types/base_types/JSONMetadata.py
@@ -1,6 +1,6 @@
 from .Media import Media
 
-from ..private._common import MEDIA_TMP
+from .._private import MEDIA_TMP
 
 import codecs
 import os
