+++ wandb/sdk/data_types/helper_types/ImageMask.py
@@ -1,5 +1,5 @@
 from ..base_types.Media import Media
-from ..private._common import MEDIA_TMP
+from .._private import MEDIA_TMP
 
 import numbers
 import os
